refactor(importer) fix missing attribute strategy configuration
