+++ module/importer/src/Infrastructure/Action/Process/Product/Strategy/ImportProductMultiSelectAttributeStrategy.php
@@ -6,7 +6,7 @@
 
 declare(strict_types=1);
 
-namespace Ergonode\Importer\Infrastructure\Action\Process\Product\Resolver;
+namespace Ergonode\Importer\Infrastructure\Action\Process\Product\Strategy;
 
 use Ergonode\SharedKernel\Domain\Aggregate\AttributeId;
 use Ergonode\Attribute\Domain\ValueObject\AttributeCode;
