+++ module/importer/src/Resources/config/services.yml
@@ -40,9 +40,9 @@ services:
         arguments:
             $targetDirectory: '%import_directory%'
 
-    Ergonode\Importer\Infrastructure\Provider\AttributeProposalProvider:
+    Ergonode\Importer\Infrastructure\Action\Process\Product\ImportProductAttributeBuilder:
         arguments:
-            - '@Ergonode\Importer\Infrastructure\Proposal\Strategy\TextAreaProposalStrategy'
-            - '@Ergonode\Importer\Infrastructure\Proposal\Strategy\NumericProposalStrategy'
-            - '@Ergonode\Importer\Infrastructure\Proposal\Strategy\MultiSelectProposalStrategy'
-            - '@Ergonode\Importer\Infrastructure\Proposal\Strategy\TextProposalStrategy'
+            $strategies:
+                - '@Ergonode\Importer\Infrastructure\Action\Process\Product\Strategy\ImportProductImageAttributeStrategy'
+                - '@Ergonode\Importer\Infrastructure\Action\Process\Product\Strategy\ImportProductMultiSelectAttributeStrategy'
+                - '@Ergonode\Importer\Infrastructure\Action\Process\Product\Strategy\ImportProductSelectAttributeStrategy'
