+++ module/importer/tests/Infrastructure/Action/Process/Product/Resolver/ImportProductSelectAttributeStrategyTest.php
@@ -8,7 +8,7 @@ declare(strict_types=1);
 
 namespace Ergonode\Importer\Tests\Infrastructure\Action\Process\Product\Resolver;
 
-use Ergonode\Importer\Infrastructure\Action\Process\Product\Resolver\ImportProductSelectAttributeStrategy;
+use Ergonode\Importer\Infrastructure\Action\Process\Product\Strategy\ImportProductSelectAttributeStrategy;
 use PHPUnit\Framework\TestCase;
 use Ergonode\Attribute\Domain\ValueObject\AttributeType;
 use Ergonode\Attribute\Domain\Entity\Attribute\SelectAttribute;
